Synchronize zh-hant and zh versions.

+++ zh-hant/codes/kotlin/chapter_dynamic_programming/min_cost_climbing_stairs_dp.kt
@@ -19,7 +19,7 @@ fun minCostClimbingStairsDP(cost: IntArray): Int {
     dp[2] = cost[2]
     // 狀態轉移：從較小子問題逐步求解較大子問題
     for (i in 3..n) {
-        dp[i] = (min(dp[i - 1].toDouble(), dp[i - 2].toDouble()) + cost[i]).toInt()
+        dp[i] = min(dp[i - 1], dp[i - 2]) + cost[i]
     }
     return dp[n]
 }
@@ -32,7 +32,7 @@ fun minCostClimbingStairsDPComp(cost: IntArray): Int {
     var b = cost[2]
     for (i in 3..n) {
         val tmp = b
-        b = (min(a.toDouble(), tmp.toDouble()) + cost[i]).toInt()
+        b = min(a, tmp) + cost[i]
         a = tmp
     }
     return b
